+++ kernel/src/pcb/pcb.h
@@ -74,7 +74,7 @@ void set_estado_pcb(t_pcb *pcb, t_state estado);
 void set_estimacion_rafaga_pcb(t_pcb *pcb, double estimacion);
 void set_rafaga_ejecutada_pcb(t_pcb *pcb, u_int64_t rafaga);
 
-int8_t es_de_menor_tamanio_que(t_pcb *proceso_a, t_pcb *proceso_b);
+int32_t es_de_menor_tamanio_que(t_pcb *proceso_a, t_pcb *proceso_b);
 
 void debug_pcb(t_pcb *pcb);
